refactor: 收敛 Core runtime 装配边界
